+++ packages/eslint-plugin-internal/tests/RuleTester.ts
@@ -1,4 +1,4 @@
-import { ESLintUtils } from '@typescript-eslint/experimental-utils';
+import { ESLintUtils } from '@typescript-eslint/utils';
 import path from 'path';
 
 function getFixturesRootDir(): string {